fix(rbac): move basic auth out of ee
