+++ packages/tracecat-ee/tracecat_ee/rbac/router.py
@@ -28,16 +28,11 @@ from tracecat_ee.rbac.schemas import (
     GroupRoleAssignmentUpdate,
     GroupUpdate,
     RoleCreate,
-    RoleList,
     RoleReadWithScopes,
     RoleUpdate,
     ScopeCreate,
     ScopeList,
     ScopeRead,
-    UserRoleAssignmentCreate,
-    UserRoleAssignmentList,
-    UserRoleAssignmentReadWithDetails,
-    UserRoleAssignmentUpdate,
 )
 from tracecat_ee.rbac.service import RBACService
 
@@ -164,38 +159,6 @@ roles_router = APIRouter(
     tags=["rbac"],
     dependencies=[Depends(_require_rbac_entitlement)],
 )
-
-
-@roles_router.get("", response_model=RoleList)
-@require_scope("org:rbac:read")
-async def list_roles(
-    *,
-    role: OrgUserRole,
-    session: AsyncDBSession,
-) -> RoleList:
-    """List roles for the organization.
-
-    Requires: org:rbac:read scope
-    """
-    service = RBACService(session, role=role)
-    roles = await service.list_roles()
-    return RoleList(
-        items=[
-            RoleReadWithScopes(
-                id=r.id,
-                name=r.name,
-                slug=r.slug,
-                description=r.description,
-                organization_id=r.organization_id,
-                created_at=r.created_at,
-                updated_at=r.updated_at,
-                created_by=r.created_by,
-                scopes=ScopeRead.list_adapter().validate_python(r.scopes),
-            )
-            for r in roles
-        ],
-        total=len(roles),
-    )
 
 
 @roles_router.get("/{role_id}", response_model=RoleReadWithScopes)
@@ -775,190 +738,5 @@ async def delete_assignment(
     service = RBACService(session, role=role)
     try:
         await service.delete_group_role_assignment(assignment_id)
-    except TracecatNotFoundError as e:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=str(e)) from e
-
-
-# =============================================================================
-# User Role Assignments Router
-# =============================================================================
-
-user_assignments_router = APIRouter(
-    prefix="/rbac/user-assignments",
-    tags=["rbac"],
-    dependencies=[Depends(_require_rbac_entitlement)],
-)
-
-
-@user_assignments_router.get("", response_model=UserRoleAssignmentList)
-@require_scope("org:rbac:read")
-async def list_user_assignments(
-    *,
-    role: OrgUserRole,
-    session: AsyncDBSession,
-    user_id: UUID | None = Query(None, description="Filter by user ID"),
-    workspace_id: UUID | None = Query(None, description="Filter by workspace ID"),
-) -> UserRoleAssignmentList:
-    """List user role assignments for the organization.
-
-    Requires: org:rbac:read scope
-    """
-    service = RBACService(session, role=role)
-    assignments = await service.list_user_assignments(
-        user_id=user_id,
-        workspace_id=workspace_id,
-    )
-    return UserRoleAssignmentList(
-        items=[
-            UserRoleAssignmentReadWithDetails(
-                id=a.id,
-                organization_id=a.organization_id,
-                user_id=a.user_id,
-                workspace_id=a.workspace_id,
-                role_id=a.role_id,
-                assigned_at=a.assigned_at,
-                assigned_by=a.assigned_by,
-                user_email=a.user.email,
-                role_name=a.role.name,
-                workspace_name=a.workspace.name if a.workspace else None,
-            )
-            for a in assignments
-        ],
-        total=len(assignments),
-    )
-
-
-@user_assignments_router.get(
-    "/{assignment_id}", response_model=UserRoleAssignmentReadWithDetails
-)
-@require_scope("org:rbac:read")
-async def get_user_assignment(
-    *,
-    role: OrgUserRole,
-    session: AsyncDBSession,
-    assignment_id: UUID,
-) -> UserRoleAssignmentReadWithDetails:
-    """Get a user role assignment by ID.
-
-    Requires: org:rbac:read scope
-    """
-    service = RBACService(session, role=role)
-    try:
-        a = await service.get_user_assignment(assignment_id)
-        return UserRoleAssignmentReadWithDetails(
-            id=a.id,
-            organization_id=a.organization_id,
-            user_id=a.user_id,
-            workspace_id=a.workspace_id,
-            role_id=a.role_id,
-            assigned_at=a.assigned_at,
-            assigned_by=a.assigned_by,
-            user_email=a.user.email,
-            role_name=a.role.name,
-            workspace_name=a.workspace.name if a.workspace else None,
-        )
-    except TracecatNotFoundError as e:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=str(e)) from e
-
-
-@user_assignments_router.post(
-    "",
-    response_model=UserRoleAssignmentReadWithDetails,
-    status_code=status.HTTP_201_CREATED,
-)
-@require_scope("org:rbac:create")
-async def create_user_assignment(
-    *,
-    role: OrgUserRole,
-    session: AsyncDBSession,
-    params: UserRoleAssignmentCreate,
-) -> UserRoleAssignmentReadWithDetails:
-    """Create a user role assignment.
-
-    Assigns a role directly to a user. If workspace_id is None, creates an org-wide
-    assignment that applies to all workspaces. Each user can have at most
-    one assignment per workspace (or one org-wide assignment).
-
-    Requires: org:rbac:create scope
-    """
-    service = RBACService(session, role=role)
-    try:
-        a = await service.create_user_assignment(
-            user_id=params.user_id,
-            role_id=params.role_id,
-            workspace_id=params.workspace_id,
-        )
-        return UserRoleAssignmentReadWithDetails(
-            id=a.id,
-            organization_id=a.organization_id,
-            user_id=a.user_id,
-            workspace_id=a.workspace_id,
-            role_id=a.role_id,
-            assigned_at=a.assigned_at,
-            assigned_by=a.assigned_by,
-            user_email=a.user.email,
-            role_name=a.role.name,
-            workspace_name=a.workspace.name if a.workspace else None,
-        )
-    except TracecatNotFoundError as e:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=str(e)) from e
-    except IntegrityError as e:
-        raise HTTPException(
-            status_code=status.HTTP_409_CONFLICT,
-            detail="User already has an assignment for this workspace",
-        ) from e
-
-
-@user_assignments_router.patch(
-    "/{assignment_id}", response_model=UserRoleAssignmentReadWithDetails
-)
-@require_scope("org:rbac:update")
-async def update_user_assignment(
-    *,
-    role: OrgUserRole,
-    session: AsyncDBSession,
-    assignment_id: UUID,
-    params: UserRoleAssignmentUpdate,
-) -> UserRoleAssignmentReadWithDetails:
-    """Update a user role assignment (change role).
-
-    Requires: org:rbac:update scope
-    """
-    service = RBACService(session, role=role)
-    try:
-        a = await service.update_user_assignment(assignment_id, role_id=params.role_id)
-        return UserRoleAssignmentReadWithDetails(
-            id=a.id,
-            organization_id=a.organization_id,
-            user_id=a.user_id,
-            workspace_id=a.workspace_id,
-            role_id=a.role_id,
-            assigned_at=a.assigned_at,
-            assigned_by=a.assigned_by,
-            user_email=a.user.email,
-            role_name=a.role.name,
-            workspace_name=a.workspace.name if a.workspace else None,
-        )
-    except TracecatNotFoundError as e:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=str(e)) from e
-
-
-@user_assignments_router.delete(
-    "/{assignment_id}", status_code=status.HTTP_204_NO_CONTENT
-)
-@require_scope("org:rbac:delete")
-async def delete_user_assignment(
-    *,
-    role: OrgUserRole,
-    session: AsyncDBSession,
-    assignment_id: UUID,
-) -> None:
-    """Delete a user role assignment.
-
-    Requires: org:rbac:delete scope
-    """
-    service = RBACService(session, role=role)
-    try:
-        await service.delete_user_assignment(assignment_id)
     except TracecatNotFoundError as e:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=str(e)) from e
